exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -52,8 +52,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const userid = decodedToken.id
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === userid.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).end()
+  } 
 })
 
 blogsRouter.put('/:id', async (request, response) => {
